projects/clash_of_codes_scraper/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0

# Optional: For JavaScript-heavy pages
# playwright>=1.40.0
# requests-html>=0.10.0

# Utilities
urllib3>=2.0.0
tqdm>=4.66.0  # Progress bars
